% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODclustsig.R
\docType{package}
\name{MODclustsig}
\alias{MODclustsig}
\title{A mod of the package clustsig to include the Hellinger Distance as method}
\description{
A mod of the package clustsig to include the Hellinger Distance as method
}
