% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simprof-dendrogram.R
\name{plot,simprof,missing-method}
\alias{plot,simprof,missing-method}
\title{Plot simplot objects}
\usage{
\S4method{plot}{simprof,missing}(x, y, ...)
}
\arguments{
\item{x}{simplot.}

\item{y}{missing.}

\item{...}{other arguments for simprof.plot.}
}
\value{
A dendrogram.
}
\description{
Method to plot a simplot object. It calls \code{simprof.plot}
}
\examples{
NULL
}
\seealso{
\code{\link{simprof.plot}}
}
