% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{simprof-class}
\alias{simprof-class}
\title{simprof_class}
\value{
S4 object.
}
\description{
simprof_class
}
\section{Slots}{

\describe{
\item{\code{numgroups}}{The number of groups which are found to bestatistically significant..}

\item{\code{significantclusters}}{A list of length numgroups with each element containing the sample IDs (row/column numbers in the corresponding original data) that are in each significant cluster.}

\item{\code{pval}}{The merge component from the hclust results with an extra column of p-values. These p-values are for testing whether the two groups in that row are statistically different.}

\item{\code{hclust}}{The S3 object returned by hclust.}
}}

\examples{
NULL
}
